[Anggi] Fixing Coin Response

+++ app/src/main/java/com/exercise/ministockbitapp/data/dto/CoinRawPrice.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class CoinRawPrice(
         @SerializedName("CHANGE24HOUR")
-        val lastPrice24Hour: Integer,
+        val lastPrice24Hour: Double,
         @SerializedName("CHANGEPCT24HOUR")
-        val lastPrice24HourPct: Integer
+        val lastPrice24HourPct: Double
 )
